backend/requirements.txt
accelerate==1.3.0
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
bitsandbytes==0.45.1
blinker==1.9.0
blobfile==3.0.0
celery==5.5.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
distlib==0.3.9
distro==1.9.0
Django==5.1.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
drf-spectacular==0.27.2
PyJWT==2.9.0
python-jose==3.3.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
filelock==3.16.1
fire==0.7.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.1.1
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.5.4
llama_models==0.0.61
llama_stack==0.0.61
llama_stack_client==0.0.61
lxml==5.3.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==2.2.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
pyaml==24.12.1
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
pygame==2.6.1
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
scipy==1.15.1
selenium==4.25.0
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.41
sqlparse==0.5.1
sympy==1.13.1
termcolor==2.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.48.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
vine==5.1.0
virtualenv==20.28.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
whitenoise==6.9.0
wsproto==1.2.0
playwright==1.48.0
dj-database-url==2.2.0
pytesseract==0.3.13
rapidfuzz==3.9.3
haversine==2.8.1
